+++ spec/sunburst-chart-spec.js
@@ -1,4 +1,4 @@
-/* global appendChartID, loadDateFixture, makeDate */
+/* global appendChartID, loadDateFixture, makeDate, getSunburstDataOneRing3Segments, loadSunburstData3CompleteRings */
 describe('dc.sunburstChart', () => {
     const width = 200;
     const height = 200;
@@ -293,6 +293,155 @@ describe('dc.sunburstChart', () => {
         afterEach(() => {
             dateDimension.filterAll();
         });
+    });
+
+    describe('sunburst use baseMixin.ordering', () => {
+        function buildSunburstChartOneRingThreeSlices (id) {
+            data = crossfilter(getSunburstDataOneRing3Segments());
+            const _valueDimension = data.dimension(d => [d.x]);
+            valueGroup = _valueDimension.group().reduceSum(d => +d.y);
+            appendChartID(id);
+            const chart = dc.sunburstChart(`#${id}`);
+            chart
+                .dimension(_valueDimension)
+                .group(valueGroup)
+                .width(width)
+                .height(height)
+                .transitionDuration(0);
+            return chart;
+        }
+
+        function expectTextLabels (strings) {
+            strings.forEach((str,i) => {
+                expect(d3.select(`text.pie-slice._${i}`).text()).toEqual(str);
+            });
+        }
+
+        let chart;
+        beforeEach(() => {
+            chart = buildSunburstChartOneRingThreeSlices('sunburst_ordering_default_ordering');
+            chart.render();
+        });
+
+        describe('sunburst using default ordering', () => {
+            it('slices ordered by key', () => {
+                expectTextLabels(['a', 'b', 'c']);
+            });
+        });
+
+        describe('sunburst using ordering by value ascending', () => {
+            it('slices ordered by value', () => {
+                chart.ordering(d => -d.value);
+                chart.render();
+                expectTextLabels(['c', 'b', 'a']);
+            });
+        });
+
+    });
+
+    describe('sunburst.relativeRingSizes', () => {
+
+        function buildSunburstChart3CompleteRings (id) {
+            data = crossfilter(loadSunburstData3CompleteRings());
+            const _valueDimension = data.dimension(d => [d.x1, d.x2, d.x3]);
+            valueGroup = _valueDimension.group().reduceSum(d => +d.y);
+            appendChartID(id);
+            const chart = dc.sunburstChart(`#${id}`);
+            chart
+                .dimension(_valueDimension)
+                .group(valueGroup)
+                .width(width)
+                .height(height)
+                .transitionDuration(0);
+            return chart;
+        }
+
+        function getPieSliceBBoxY (chart, sliceNumber) {
+            return chart.select(`.pie-slice._${sliceNumber}`).node().getBBox().y;
+        }
+
+        function getRingThicknessRounded (chart, ringNumber) {
+            if (ringNumber === 0) {
+                throw new Error('root ring 0 can not be checked this way.');
+            }
+            const yInner = getPieSliceBBoxY(chart, ringNumber - 1);
+            const yOuter = getPieSliceBBoxY(chart, ringNumber);
+            return Math.round(Math.abs(yOuter - yInner));
+        }
+
+        describe('sunburst.relativeRingSizes regression', () => {
+            let chart;
+            beforeEach(() => {
+                chart = buildSunburstChart3CompleteRings('sunburst_relativeRingSizes_regression');
+                chart.render();
+            });
+
+            it('rings should get narrower, farther away from the center', () => {
+                expect(getRingThicknessRounded(chart, 1)).toBeGreaterThan(getRingThicknessRounded(chart, 2));
+            });
+        });
+
+        describe('sunburst.relativeRingSizes: equal distribution', () => {
+            let chart;
+            beforeEach(() => {
+                chart = buildSunburstChart3CompleteRings('sunburst_relativeRingSizes_equal_distribution');
+                chart.ringSizes(chart.equalRingSizes());
+                chart.render();
+            });
+            it('rings should be equally wide', () => {
+                expect(getRingThicknessRounded(chart, 1)).toEqual(getRingThicknessRounded(chart, 2));
+            });
+        });
+
+        describe('sunburst.relativeRingSizes: specific percentages', () => {
+            let chart;
+            const specificPercentages = function (ringCount) {
+                return [.1, .3, .6];
+            };
+            beforeEach(() => {
+                chart = buildSunburstChart3CompleteRings('sunburst_relativeRingSizes_specific_percentages');
+                chart.ringSizes(chart.relativeRingSizes(specificPercentages));
+                chart.render();
+            });
+            it('2nd ring should be half as wide as the 3rd ', () => {
+                expect(2 * getRingThicknessRounded(chart, 1)).toEqual(getRingThicknessRounded(chart, 2));
+            });
+        });
+
+        describe('sunburst.relativeRingSizes: invalid arguments', () => {
+            let chart;
+
+            const functionReturnsNonArray = function (ringCount) {
+                return {};
+            };
+
+            const tooManyPercentageValues = function (ringCount) {
+                return [.1, .1, .1, .1];
+            };
+
+            const percentagesSumNot1 = function (ringCount) {
+                return [.5, .5, .5];
+            };
+
+            beforeEach(() => {
+                chart = buildSunburstChart3CompleteRings('sunburst_relativeRingSizes_invalid_arguments');
+            });
+
+            it('invalid arguments cause dc.errors.BadArgumentException, default function does not', () => {
+                chart.ringSizes(chart.relativeRingSizes(functionReturnsNonArray));
+                expect(() => {chart.render()}).toThrowError(dc.BadArgumentException);
+
+                chart.ringSizes(chart.relativeRingSizes(tooManyPercentageValues));
+                expect(() => {chart.render()}).toThrowError(dc.BadArgumentException);
+
+                chart.ringSizes(chart.relativeRingSizes(percentagesSumNot1));
+                expect(() => {chart.render()}).toThrowError(dc.BadArgumentException);
+
+                chart.ringSizes(chart.defaultRingSizes());
+                chart.render();
+            });
+        });
+
     });
 
 });